New backend, full parallelism, improved I/O (#21)
* cleaned up plotting
* added set/get_node_attribute
* added types for get_degrees
* type checks for float/int + when loading from string
* fixed syn_spec in prop and examples testing
* updated MANIFEST, improved cython cleanup, added example, improved setup install
* corrected compile option, added global lambertw for test + reset_kernel and removed nest
* fix scipy install on travis for py3.5
* new default graph backend
* parallel MPI saving
* full shape and population saving in graph I/O
* updated graph-tool install
* improved random seeding
* updated doc
* improved get_spikes

+++ nngt/generation/cconnect.pxd
@@ -16,7 +16,7 @@ from libcpp cimport bool
 cdef extern from "func_connect.h" namespace "generation":
     cdef void _gen_edges(
       size_t* ia_edges, const vector[size_t]& first_nodes,
-      const vector[size_t]& degrees, const vector[size_t]& second_nodes,
+      const vector[unsigned int]& degrees, const vector[size_t]& second_nodes,
       const vector[ vector[size_t] ]& existing_edges, unsigned int idx,
       bool multigraph, bool directed, long msd, unsigned int omp) except +
